chore: rename backup wallet screen files
