+++ components/brave_wallet_ui/page/screens/onboarding/onboarding.routes.tsx
@@ -12,7 +12,7 @@ import {
 } from 'react-router'
 
 // components
-import { OnboardingBackupRecoveryPhrase } from '../backup-wallet/backup-recovery-phrase/onboarding-backup-recovery-phrase'
+import { OnboardingBackupRecoveryPhrase } from '../backup-wallet/backup-recovery-phrase/backup-recovery-phrase'
 import { OnboardingCreatePassword } from './create-password/onboarding-create-password'
 import { OnboardingRecoveryPhraseExplainer } from '../backup-wallet/explain-recovery-phrase/explain-recovery-phrase'
 import { OnboardingVerifyRecoveryPhrase } from '../backup-wallet/verify-recovery-phrase/verify-recovery-phrase'
